Project/Senior-Project-main/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
bcrypt==4.0.1
python-dotenv==1.0.0